Implement Ignore Patterns
Some pages have overbearing cookie messages. Ignore patterns is a
list of regexes that the extractor will use when identifying
text nodes. Any nodes whose text matches at least one of the regexes
will be ignored.

NB this does not guarantee that the ignored node will not appear in the
parsed text. Even if a particular node is ignored and not scored, it might
be related to another node that scores highly.

The rationale for this feature was to handle pages with overbearing
cookie messages.